Complete from meta table via "transforming" grammars
This involves a slight rethinking of how the meta
grammar works. Since it's primarily the place where grammar
definitions and (correspondingly) completion suggestions are stored,
we add functions for "transforming" these garmmars when they are copied
over into the root grammar (where the user can directly manipulate).

This transformation happens in the process of rendering the grammar in
the view, keeping the Grammar representation the same within the model.

This is a compromise between previous implementation attempts that either
treat grammar definition as a totally native grammar, or as an exact
mirror of their representation within root.